// apps/browser-extension/src/common/types.ts
export enum SourceKey {
  URL = 'Url',
  HOST = 'host',
  PATH = 'path',
}

export enum SourceOperator {
  EQUALS = 'Equals',
  CONTAINS = 'Contains',
  MATCHES = 'Matches',
  WILDCARD_MATCHES = 'Wildcard_Matches',
}

export enum SourceFilterTypes {
  PAGE_DOMAINS = 'pageDomains',
  REQUEST_METHOD = 'requestMethod',
  RESOURCE_TYPE = 'resourceType',
  REQUEST_PAYLOAD = 'requestPayload',
}

export interface UrlSource {
  id?: string;
  isActive?: boolean;
  key: SourceKey;
  operator: SourceOperator;
  value: string;
  filters?: RuleSourceFilter[];
}

export interface RuleSourceFilter {
  pageDomains: string[];
  requestMethod: HttpRequestMethod[];
  resourceType: ResourceType[];
  requestPayload?: RequestPayloadFilter;
}

export interface RequestPayloadFilter {
  key: string;
  operator?: 'Contains' | 'Equals';
  value: string;
}

export enum ResourceType {
  XHR = 'xmlhttprequest',
  JS = 'script',
  CSS = 'stylesheet',
  Image = 'image',
  Media = 'media',
  Font = 'font',
  WebSocket = 'websocket',
  MainDocument = 'main_frame',
  IFrameDocument = 'sub_frame',
}

enum HttpRequestMethod {
  GET = 'GET',
  POST = 'POST',
  PUT = 'PUT',
  DELETE = 'DELETE',
  PATCH = 'PATCH',
  OPTIONS = 'OPTIONS',
  CONNECT = 'CONNECT',
  HEAD = 'HEAD',
}